Added findbugs and checkstyle. Fixed some basic findbugs issues
